refactor: extract FadeSlider component
